Move sortable into partners namespaced helper
- Also bring back helper_methods I removed in children_controller, it
turns out they are being used by this sortable method
